requirements.txt
torch
torchvision
transformers
pillow
kafka-python
timm
